location: split LR1110 and nRF70 Wi-Fi PAL implementations
Extract LR1110 Wi-Fi PAL into a dedicated file and link it only when
nRF70 Wi-Fi location is disabled. Add a separate GNSS stub for
nRF70-only builds and skip sx126x location stubs when the nRF70 PAL is
active.
