+++ subsys/semtech/CMakeLists.txt
@@ -68,9 +68,6 @@ target_compile_definitions(smtc_lbm INTERFACE
 
 add_subdirectory_ifdef(CONFIG_SIDEWALK_SUBGHZ_RADIO_SX126X sx126x)
 
-# When the nRF70 Wi-Fi PAL is active it provides the GNSS/Wi-Fi PAL symbols
-# instead of the sx126x_location NOSUPPORT stub, so skip the stub to avoid
-# duplicate symbol definitions.
 if(CONFIG_SIDEWALK_SUBGHZ_RADIO_SX126X AND NOT CONFIG_SIDEWALK_WIFI_LOCATION_NRF70)
 	add_subdirectory(sx126x_location)
 endif()
@@ -85,9 +82,13 @@ if(CONFIG_SIDEWALK_SUBGHZ_RADIO_SX126X)
 	target_link_libraries(sidewalk_radio INTERFACE
 		"-Wl,--start-group"
 		sid_pal_radio_sx126x_impl
-		sid_pal_radio_sx126x_location_impl
 		"-Wl,--end-group"
 	)
+	if(NOT CONFIG_SIDEWALK_WIFI_LOCATION_NRF70)
+		target_link_libraries(sidewalk_radio INTERFACE
+			sid_pal_radio_sx126x_location_impl
+		)
+	endif()
 endif()
 
 if(CONFIG_SIDEWALK_SUBGHZ_RADIO_LR1110)
